mix precision training

+++ modeling/llm_post_training/instruction_sft.py
@@ -106,7 +106,7 @@ def main(args):
 
     model = AutoModelForCausalLM.from_pretrained(
         model_name,
-        torch_dtype=(torch.float16 if torch.cuda.is_available() else torch.float32),
+        torch_dtype=(torch.bfloat16 if torch.cuda.is_available() else torch.float32),
         device_map="auto" if torch.cuda.is_available() else None,
         trust_remote_code=True,
     )
